CloudNet and Redis Intergration
Redis Intergration has entered it's beta stages along with the introduction of CloudNet v3 into the server.

So far when booting the server makes keys in redis (for server settings) that will probably be expanded at a later date.

A know bug at the moment is that on server shutdown the server is not removed form the online servers list.